Split out edit view into a separate module
