+++ src/main.rs
@@ -34,9 +34,10 @@ extern crate xi_core_lib;
 extern crate xi_rpc;
 extern crate xi_win_shell;
 
+mod dialog;
+mod edit_view;
 mod linecache;
 mod menus;
-mod dialog;
 mod xi_thread;
 
 use std::cell::RefCell;
@@ -44,113 +45,36 @@ use std::sync::mpsc::TryRecvError;
 use std::rc::Rc;
 
 use winapi::shared::windef::*;
-use winapi::um::winuser::*;
-
-use direct2d::brush;
-use direct2d::math::*;
-use directwrite::text_format::{self, TextFormat};
-use directwrite::text_layout::{self, TextLayout};
 
 use serde_json::Value;
 
-use linecache::LineCache;
+use edit_view::EditView;
 use menus::MenuEntries;
 use xi_win_shell::util::Error;
 use dialog::{get_open_file_dialog_path, get_save_file_dialog_path};
 use xi_thread::{start_xi_thread, XiPeer};
 
 use xi_win_shell::paint::PaintCtx;
-use xi_win_shell::util::default_text_options;
 use xi_win_shell::win_main::{self, RunLoopHandle};
 use xi_win_shell::window::{WindowBuilder, WindowHandle, WinHandler};
 
-struct Resources {
-    fg: brush::SolidColor,
-    bg: brush::SolidColor,
-    text_format: TextFormat,
-}
-
-impl Resources {
-    fn create_text_layout(&self, factory: &directwrite::Factory, text: &str) -> TextLayout {
-        let params = text_layout::ParamBuilder::new()
-            .text(text)
-            .font(self.text_format.clone())
-            .width(1e6)
-            .height(1e6)
-            .build().unwrap();
-        factory.create(params).unwrap()
-    }
-}
-
 struct MainWinState {
     rpc_id: usize,
-    view_id: String,
-    line_cache: LineCache,
     label: String,
-    dwrite_factory: directwrite::Factory,
-    resources: Option<Resources>,
-    filename: Option<String>,
+    edit_view: EditView,
 }
 
 impl MainWinState {
     fn new() -> MainWinState {
         MainWinState {
             rpc_id: 0,
-            view_id: String::new(),
-            line_cache: LineCache::new(),
             label: "hello direct2d".to_string(),
-            dwrite_factory: directwrite::Factory::new().unwrap(),
-            resources: None,
-            filename: None
-        }
-    }
-
-    fn create_resources(&mut self, p: &mut PaintCtx) -> Resources {
-        let rt = p.render_target();
-        let text_format_params = text_format::ParamBuilder::new()
-            .size(15.0)
-            .family("Consolas")
-            .build().unwrap();
-        let text_format = self.dwrite_factory.create(text_format_params).unwrap();
-        Resources {
-            fg: rt.create_solid_color_brush(0xf0f0ea, &BrushProperties::default()).unwrap(),
-            bg: rt.create_solid_color_brush(0x272822, &BrushProperties::default()).unwrap(),
-            text_format: text_format,
+            edit_view: EditView::new(),
         }
     }
 
     fn render(&mut self, p: &mut PaintCtx) {
-        if self.resources.is_none() {
-            self.resources = Some(self.create_resources(p));
-        }
-        let resources = &self.resources.as_ref().unwrap();
-        let rt = p.render_target();
-        let size = rt.get_size();
-        let rect = RectF::from((0.0, 0.0, size.width, size.height));
-        rt.fill_rectangle(&rect, &resources.bg);
-
-        let x0 = 6.0;
-        let mut y = 6.0;
-        for line_num in 0..self.line_cache.height() {
-            if let Some(line) = self.line_cache.get_line(line_num) {
-                let layout = resources.create_text_layout(&self.dwrite_factory, line.text());
-                rt.draw_text_layout(
-                    &Point2F::from((x0, y)),
-                    &layout,
-                    &resources.fg,
-                    default_text_options()
-                );
-                for &offset in line.cursor() {
-                    if let Some(pos) = layout.hit_test_text_position(offset as u32, true) {
-                        let x = x0 + pos.point_x;
-                        rt.draw_line(&Point2F::from((x, y)),
-                            &Point2F::from((x, y + 17.0)),
-                            &resources.fg, 1.0, None);
-                    }
-                }
-            }
-            y += 17.0;
-        }
+        self.edit_view.render(p);
     }
 }
 
@@ -159,7 +83,7 @@ struct MainWinHandler {
 }
 
 // Maybe combine all this, put as a single item inside a RefCell.
-struct MainWin {
+pub struct MainWin {
     peer: XiPeer,
     handle: RefCell<WindowHandle>,
     state: RefCell<MainWinState>,
@@ -183,8 +107,7 @@ impl MainWin {
     }
 
     // Note: caller can't be borrowing the state.
-    fn send_edit_cmd(&self, method: &str, params: &Value) {
-        let view_id = &self.state.borrow_mut().view_id;
+    fn send_edit_cmd(&self, method: &str, params: &Value, view_id: &str) {
         let edit_params = json!({
             "method": method,
             "params": params,
@@ -198,19 +121,19 @@ impl MainWin {
         if let Some(filename) = filename {
             self.req_new_view(Some(&filename));
             let mut state = self.state.borrow_mut();
-            state.filename = Some(filename);
-            state.line_cache = LineCache::new();
+            state.edit_view.filename = Some(filename);
+            state.edit_view.clear_line_cache();
         }
     }
 
     fn file_save(&self, hwnd_owner: HWND) {
-        let filename: Option<String> = self.state.borrow_mut().filename.clone();
+        let filename: Option<String> = self.state.borrow_mut().edit_view.filename.clone();
         if filename.is_none() {
             self.file_save_as(hwnd_owner);
         } else {
             let state = self.state.borrow_mut();
             self.send_notification("save", &json!({
-                "view_id": state.view_id,
+                "view_id": state.edit_view.view_id,
                 "file_path": filename,
             }));
         }
@@ -220,11 +143,11 @@ impl MainWin {
         if let Some(filename) = unsafe { get_save_file_dialog_path(hwnd_owner) } {
             let mut state = self.state.borrow_mut();
             self.send_notification("save", &json!({
-                "view_id": state.view_id,
+                "view_id": state.edit_view.view_id,
                 "file_path": filename,
             }));
 
-            state.filename = Some(filename.clone());
+            state.edit_view.filename = Some(filename.clone());
         }
     }
 
@@ -277,44 +200,14 @@ impl WinHandler for MainWinHandler {
         }
     }
 
-    fn char(&self, ch: u32, _mods: u32) {
-        match ch {
-            0x08 => {
-                self.win.send_edit_cmd("delete_backward", &json!([]));
-            },
-            0x0d => {
-                self.win.send_edit_cmd("insert_newline", &json!([]));
-            },
-            _ => {
-                if let Some(c) = ::std::char::from_u32(ch) {
-                    let params = json!({"chars": c.to_string()});
-                    self.win.send_edit_cmd("insert", &params);
-                }
-            }
-        }
+    fn char(&self, ch: u32, mods: u32) {
+        let edit_view = &mut self.win.state.borrow_mut().edit_view;
+        edit_view.char(ch, mods, &self.win);
     }
 
-    fn keydown(&self, vk_code: i32, _mods: u32) -> bool {
-        // Handle special keys here
-        match vk_code {
-            VK_UP => {
-                self.win.send_edit_cmd("move_up", &json!([]));
-            },
-            VK_DOWN => {
-                self.win.send_edit_cmd("move_down", &json!([]));
-            },
-            VK_LEFT => {
-                self.win.send_edit_cmd("move_left", &json!([]));
-            },
-            VK_RIGHT => {
-                self.win.send_edit_cmd("move_right", &json!([]));
-            },
-            VK_DELETE => {
-                self.win.send_edit_cmd("delete_forward", &json!([]));
-            },
-            _ => return false
-        }
-        true
+    fn keydown(&self, vk_code: i32, mods: u32) -> bool {
+        let edit_view = &mut self.win.state.borrow_mut().edit_view;
+        edit_view.keydown(vk_code, mods, &self.win)
     }
 
     fn destroy(&self) {
@@ -328,11 +221,11 @@ impl MainWin {
         //println!("got {:?}", v);
         if let Some(tab_name) = v["result"].as_str() {
             // TODO: should match up id etc. This is quick and dirty.
-            state.view_id = tab_name.to_string();
+            state.edit_view.set_view_id(tab_name);
         } else {
             let ref method = v["method"];
             if method == "update" {
-                state.line_cache.apply_update(&v["params"]["update"]);
+                state.edit_view.apply_update(&v["params"]["update"]);
             }
         }
         state.label = serde_json::to_string(v).unwrap();
